requirements.txt
flask
gunicorn
rembg
pillow
onnxruntime